+++ packages/clients/src/utils/TextUtils.ts
@@ -1,27 +1,7 @@
-const ROMAN_NUMBERS = [
-  '0',
-  'I',
-  'II',
-  'III',
-  'IV',
-  'V',
-  'VI',
-  'VII',
-  'VIII',
-  'IX',
-  'X',
-];
+import { toRomanDifficulty } from '@whitewater-guide/commons';
 
 const SEPARATOR = (0.5).toString()[1];
 const OTHER_SEPARATOR = SEPARATOR === ',' ? /\./ : /,/;
-
-export function toRomanDifficulty(decimalDifficulty: number): string {
-  if (Number.isInteger(decimalDifficulty)) {
-    return ROMAN_NUMBERS[decimalDifficulty];
-  }
-  const floor = Math.floor(decimalDifficulty);
-  return `${ROMAN_NUMBERS[floor]} - ${ROMAN_NUMBERS[floor + 1]}`;
-}
 
 export interface DifficultyFragment {
   difficulty: number;