+++ packages/mobile/__mocks__/react-native-fbsdk.ts
@@ -1,4 +1,9 @@
 export const LoginManager = {
   setLoginBehavior: jest.fn(),
   logOut: jest.fn(),
+  logInWithReadPermissions: jest.fn(),
+};
+
+export const AccessToken = {
+  getCurrentAccessToken: jest.fn(),
 };