+++ packages/commons/CHANGELOG.md
@@ -1,3 +1,5 @@
+## [0.0.137](https://github.com/doomsower/whitewater/compare/@whitewater-guide/commons@0.0.136...@whitewater-guide/commons@0.0.137) (2019-06-09)
+
 ## [0.0.136](https://github.com/doomsower/whitewater/compare/@whitewater-guide/commons@0.0.134...@whitewater-guide/commons@0.0.136) (2019-06-08)
 
 ## [0.0.135](https://github.com/doomsower/whitewater/compare/@whitewater-guide/commons@0.0.134...@whitewater-guide/commons@0.0.135) (2019-06-07)