+++ services/boompromo/CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.0.154](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.152...@whitewater-guide/boompromo@0.0.154) (2019-06-11)
+
+## [0.0.153](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.152...@whitewater-guide/boompromo@0.0.153) (2019-06-10)
+
+## [0.0.152](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.152...@whitewater-guide/boompromo@0.0.152) (2019-06-09)
+
+## [0.0.152](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.151...@whitewater-guide/boompromo@0.0.152) (2019-06-09)
+
 ## [0.0.151](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.150...@whitewater-guide/boompromo@0.0.151) (2019-06-09)
 
 ## [0.0.150](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.149...@whitewater-guide/boompromo@0.0.150) (2019-06-08)